+++ src/typing_ex/view/page.clj
@@ -171,7 +171,7 @@
    data はソーティング済みの[[login days] ...]"
   [self data]
   (page
-   [:h2 "Typing: 10 回以上練習した日数"]
+   [:h2 "Typing: 30 回以上練習した日数"]
    (headline 7)
    [:div {:style "margin-left:1rem;"}
     [:p "毎日ちょっとずつが伸びる秘訣。"]